+++ Src/SOS.Process/Processors/Artportalen/ArtportalenObservationFactory.cs
@@ -105,10 +105,11 @@ namespace SOS.Process.Processors.Artportalen
                 obs.Modified = verbatimObservation.EditDate.ToUniversalTime();
                 obs.OwnerInstitutionCode = verbatimObservation.OwnerOrganization?.Translate(Cultures.en_GB, Cultures.sv_SE) ?? "Artdatabanken";
                 obs.PrivateCollection = verbatimObservation.PrivateCollection;
+                obs.Projects = verbatimObservation.Projects?.Select(CreateProcessedProject);
                 obs.PublicCollection = verbatimObservation.PublicCollection?.Translate(Cultures.en_GB, Cultures.sv_SE);
                 obs.RightsHolder = verbatimObservation.RightsHolder ?? verbatimObservation.OwnerOrganization?.Translate(Cultures.en_GB, Cultures.sv_SE) ?? "Data saknas";
                 obs.Type = null;
-                
+
                 // Event
                 obs.Event = new Event();
                 obs.Event.DiscoveryMethod = GetSosIdFromMetadata(verbatimObservation?.DiscoveryMethod, VocabularyId.DiscoveryMethod);
@@ -244,7 +245,6 @@ namespace SOS.Process.Processors.Artportalen
                 obs.ArtportalenInternal.ReportedByUserAlias = verbatimObservation.ReportedByUserAlias;
                 obs.ArtportalenInternal.LocationPresentationNameParishRegion = verbatimObservation.Site?.PresentationNameParishRegion;
                 obs.ArtportalenInternal.OccurrenceRecordedByInternal = verbatimObservation.ObserversInternal;
-                obs.ArtportalenInternal.Projects = verbatimObservation.Projects?.Select(CreateProcessedProject);
                 obs.ArtportalenInternal.IncrementalHarvested = _incrementalMode;
 
                 // Set dependent properties