+++ Src/SOS.lib/Models/Processed/Observation/ArtportalenInternal.cs
@@ -118,11 +118,5 @@ namespace SOS.Lib.Models.Processed.Observation
         /// True if sighting was incremental harvested
         /// </summary>
         public bool IncrementalHarvested { get; set; }
-
-        /// <summary>
-        ///     Projects connected to sighting
-        /// </summary>
-        [Nested]
-        public IEnumerable<Project> Projects { get; set; }
     }
 }